backend/pyproject.toml
[project]
name = "task_manager_backend"
version = "0.1.0"
description = "Task Manager Kanban API"
readme = "README.md"
requires-python = ">=3.13"
authors = [{ name = "Task Manager" }]
dependencies = [
    "Django>=5.1",
    "djangorestframework>=3.14",
    "drf-spectacular>=0.27",
    "drf-spectacular-sidecar>=2024.6.1",
    "psycopg[binary]>=3.1",
    "dj-database-url>=2.1",
    "django-filter>=24.3",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.6.9",
    "mypy>=1.12.0",
    "pytest>=8.3",
    "pytest-django>=4.9",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
pythonpath = ["src"]
testpaths = ["backend/tests", "tests"]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.mypy]
python_version = "3.13"
ignore_missing_imports = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
no_implicit_optional = true

